+++ modules/tensor_mechanics/include/utils/RankFourTensor.h
@@ -1,13 +1,3 @@
-/**
- * RankFourTensor is designed to handle any fourth order tensor.
- * It is designed to allow for maximum clarity of the mathematics and ease of use.
- * Original class authors: A. M. Jokisaari, O. Heinonen, M.R. Tonks
- *
- * RankFourTensor holds the 81 separate C_ijkl entries; the entries are accessed by index, with
- * i, j, k, and l equal to 0, 1, 2
- *
- */
-
 #ifndef RANKFOURTENSOR_H
 #define RANKFOURTENSOR_H
 
@@ -27,14 +17,130 @@
 #include <vector>
 
 
+/**
+ * RankFourTensor is designed to handle any N-dimensional fourth order tensor, C.
+ *
+ * It is designed to allow for maximum clarity of the mathematics and ease of use.
+ * Original class authors: A. M. Jokisaari, O. Heinonen, M.R. Tonks
+ *
+ * Since N is hard-coded to 3, RankFourTensor holds 81 separate C_ijkl entries.
+ * Within the code i = 0, 1, 2, but this object provides methods to extract the entries
+ * with i = 1, 2, 3, and some of the documentation is also written in this way.
+ */
 class RankFourTensor
 {
 public:
+
+  /// Default constructor; fills to zero
+  RankFourTensor();
+
+  /// Copy constructor
+  RankFourTensor(const RankFourTensor &a);
+
+  /// Destructor
+  ~RankFourTensor() {}
+
+  /// Gets the value for the index specified.  Takes index = 0,1,2
+  Real & operator()(unsigned int i, unsigned int j, unsigned int k, unsigned int l);
+
+
   /**
-   * Static method for use in validParams for getting the "fill_method"
+   * Gets the value for the index specified.  Takes index = 0,1,2
+   * used for const
    */
+  Real operator()(unsigned int i, unsigned int j, unsigned int k, unsigned int l) const;
+
+  ///Sets the value for the index specified.  Indices = 0, 1, 2
+  void setValue(Real val, unsigned int i, unsigned int j, unsigned int k, unsigned int l);
+
+  /// Gets the value for the index specified.  Takes index = 1,2,3
+  Real getValue(unsigned int i, unsigned int j, unsigned int k, unsigned int l) const;
+
+  /// Zeros out the tensor.
+  void zero();
+
+  /// copies values from a into this tensor
+  RankFourTensor & operator=(const RankFourTensor &a);
+
+  /// C_ijkl*a_kl
+  RankTwoTensor operator*(const RankTwoTensor &a);
+
+  /// C_ijkl*a_kl
+  RealTensorValue operator*(const RealTensorValue &a);
+
+  /// C_ijkl*a
+  RankFourTensor operator*(const Real &a);
+
+  /// C_ijkl *= a
+  RankFourTensor & operator*=(const Real &a);
+
+  /// C_ijkl/a
+  RankFourTensor operator/(const Real &a);
+
+  /// C_ijkl /= a  for all i, j, k, l
+  RankFourTensor & operator/=(const Real &a);
+
+  /// C_ijkl += a_ijkl  for all i, j, k, l
+  RankFourTensor & operator+=(const RankFourTensor &a);
+
+  /// C_ijkl + a_ijkl
+  RankFourTensor operator+(const RankFourTensor &a) const;
+
+  /// C_ijkl -= a_ijkl
+  RankFourTensor & operator-=(const RankFourTensor &a);
+
+  /// C_ijkl - a_ijkl
+  RankFourTensor operator-(const RankFourTensor &a) const;
+
+  /// -C_ijkl
+  RankFourTensor operator - () const;
+
+  /// C_ijpq*a_pqkl
+  RankFourTensor operator*(const RankFourTensor &a) const;
+
+  /**
+   * This returns A_ijkl such that C_ijkl*A_klmn = 0.5*(de_im de_jn + de_in de_jm)
+   * This routine assumes that C_ijkl = C_jikl = C_ijlk
+   */
+  RankFourTensor invSymm();
+
+  /**
+   * Rotate the tensor using
+   * C_ijkl = R_im R_in R_ko R_lp C_mnop
+   */
+  virtual void rotate(RealTensorValue &R);
+
+  /// Print tensor using nice formatting and Moose::out
+  void print();
+
+  /**
+   * Transpose the tensor by swapping the first pair with the second pair of indices
+   * @return C_klji
+   */
+  RankFourTensor transposeMajor();
+
+
+  /**
+   * Fills the tensor entries ignoring the last dimension (ie, C_ijkl=0 if any of i, j, k, or l = 3).
+   * Fill method depends on size of input
+   * Input size = 2.  Then C_1111 = C_2222 = input[0], and C_1122 = input[1], and C_1212 = (input[0] + input[1])/2,
+                      and C_ijkl = C_jikl = C_ijlk = C_klij, and C_1211 = C_1222 = 0.
+   * Input size = 9.  Then C_1111 = input[0], C_1112 = input[1], C_1122 = input[3],
+                           C_1212 = input[4], C_1222 = input[5], C_1211 = input[6]
+                           C_2211 = input[7], C_2212 = input[8], C_2222 = input[9]
+                           and C_ijkl = C_jikl = C_ijlk
+  */
+  virtual void surfaceFillFromInputVector(const std::vector<Real> input);
+
+
+  /// Static method for use in validParams for getting the "fill_method"
   static MooseEnum fillMethodEnum();
 
+  /**
+   * To fill up the 81 entries in the 4th-order tensor, fillFromInputVector
+   * is called with one of the following fill_methods.
+   * See the fill*FromInputVector functions for more details
+   */
   enum FillMethod
   {
     antisymmetric,
@@ -46,87 +152,38 @@ public:
     general
   };
 
-  /**
-   * Default constructor; fills to zero
-   */
-  RankFourTensor();
 
   /**
-   * Copy constructor
-   */
-  RankFourTensor(const RankFourTensor &a);
-
-  ~RankFourTensor() {}
-
-  /**
-   * Gets the value for the index specified.  Takes index = 0,1,2
-   */
-  Real & operator()(unsigned int i, unsigned int j, unsigned int k, unsigned int l);
-
-
-  /**
-   * Gets the value for the index specified.  Takes index = 0,1,2,
-   * used for const
-   */
-  Real operator()(unsigned int i, unsigned int j, unsigned int k, unsigned int l) const;
-
-  /**
-   * Sets the value for the index specified
-   */
-  void setValue(Real val, unsigned int i, unsigned int j, unsigned int k, unsigned int l);
-
-  /**
-   * Gets the value for the index specified.  Takes index = 1,2,3
-   */
-  Real getValue(unsigned int i, unsigned int j, unsigned int k, unsigned int l) const;
-
-  /**
-  * Zeros out the tensor.
+  * fillFromInputVector takes some number of inputs to fill
+  * the Rank-4 tensor.
+  * @param input the numbers that will be placed in the tensor
+  * @param fill_method this can be:
+               antisymmetric (use fillAntisymmetricFromInputVector)
+               symmetric9 (use fillSymmetricFromInputVector with all=false)
+               symmetric21 (use fillSymmetricFromInputVector with all=true)
+               general_isotropic (use fillGeneralIsotropicFrominputVector)
+               symmetric_isotropic (use fillSymmetricIsotropicFromInputVector)
+               antisymmetric_isotropic (use fillAntisymmetricIsotropicFromInputVector)
+               general (use fillGeneralFromInputVector)
   */
-  void zero();
+  void fillFromInputVector(const std::vector<Real> input, FillMethod fill_method);
 
-  RankFourTensor & operator=(const RankFourTensor &a);
 
-  RankTwoTensor operator*(const RankTwoTensor &a);
+protected:
 
-  RealTensorValue operator*(const RealTensorValue &a);
+  /// Dimensionality of rank-four tensor
+  static const unsigned int N = 3;
 
-  RankFourTensor operator*(const Real &a);
+  /// The values of the rank-four tensor
+  Real _vals[N][N][N][N];
 
-  RankFourTensor & operator*=(const Real &a);
-
-  RankFourTensor operator/(const Real &a);
-
-  RankFourTensor & operator/=(const Real &a);
-
-  RankFourTensor & operator+=(const RankFourTensor &a);
-
-  RankFourTensor operator+(const RankFourTensor &a) const;
-
-  RankFourTensor & operator-=(const RankFourTensor &a);
-
-  RankFourTensor operator-(const RankFourTensor &a) const;
-
-  RankFourTensor operator - () const;
-
-  RankFourTensor operator*(const RankFourTensor &a) const;//Added
-
-  RankFourTensor invSymm();//Added
-
-  virtual void rotate(RealTensorValue &R);
   /**
-   * Print the tensor
+   * Inverts the dense matrix A using LAPACK routines
+   * @param A upon input this is a row vector representing an n-by-n matrix.  Upon output it is the inverse (as a row-vector)
+   * @param n size of A
+   * @return if zero then inversion was successful.  Otherwise A contained illegal entries or was singular
    */
-  void print();
-
-  RankFourTensor transposeMajor();
-
-
-//  int MatrixInversion(double *, int, double* );//Added
-
-  int MatrixInversion(double *, int);//Added
-
-  virtual void surfaceFillFromInputVector(const std::vector<Real> input);
+  int MatrixInversion(double *A, int n);//Added
 
    /**
   * fillSymmetricFromInputVector takes either 21 (all=true) or 9 (all=false) inputs to fill in
@@ -183,30 +240,6 @@ public:
    */
   void fillGeneralFromInputVector(const std::vector<Real> input);
 
-
-  /**
-  * fillFromInputVector takes some number of inputs to fill
-  * the Rank-4 tensor.
-  * @param input the numbers that will be placed in the tensor
-  * @param fill_method this can be:
-               antisymmetric (use fillAntisymmetricFromInputVector)
-               symmetric9 (use fillSymmetricFromInputVector with all=false)
-               symmetric21 (use fillSymmetricFromInputVector with all=true)
-               general_isotropic (use fillGeneralIsotropicFrominputVector)
-               symmetric_isotropic (use fillSymmetricIsotropicFromInputVector)
-               antisymmetric_isotropic (use fillAntisymmetricIsotropicFromInputVector)
-               general (use fillGeneralFromInputVector)
-  */
-  void fillFromInputVector(const std::vector<Real> input, FillMethod fill_method);
-
-protected:
-
-/**
- * Contains the actual data for the Rank Four tensor.
- */
-  static const unsigned int N = 3;
-
-  Real _vals[N][N][N][N];
 
 private:
